Add fail safe feature via environment variable CLAMAV_FAILSAFE
This is a patch to add fail safe feature for mailfront clamav plugin.

When the ClamAV server is unavailable, this plugin will defer all mails at
SMTP delivery including authorized users until the ClamAV server is restored.

This patch will give the system administrators the flexibility to enable fail
safe for this plugin by setting the environment variable CLAMAV_FAILSAFE. If
this environment variable is set and when the ClamAV server is unavailable,
mails will be passed and warning messages will be logged.

Details can be read at http://choon.net/mailfront-plugin-clamav.php

Signed-off-by: Giam Teck Choon <giamteckchoon@gmail.com>