Add fail safe feature via environment variable SPAMD_FAILSAFE
This is a patch to add fail safe feature for mailfront SpamAssassin plugin.

When the SpamAssassin server is unavailable, this plugin will defer all mails
at SMTP delivery including authorized users until the SpamAssassin server is
restored. This patch will give the system administrators the flexibility to
enable fail safe for this plugin by setting the environment variable
SPAMD_FAILSAFE. If this environment variable is set and when the SpamAssassin
server is unavailable, mails will be passed and warning messages will be logged.

Details can be read at http://choon.net/mailfront-plugin-spamassassin.php

Signed-off-by: Giam Teck Choon <giamteckchoon@gmail.com>