getprotoenv: retries with hardcore prefix string TCP and SSL if env=0
Try to do a second getenv if env=0 with hardcore prefix TCP to name and if
return nothing then try the third getenv with prefix SSL to name.

This is needed if proto is not "TCP" string if we are using patched ucspi-tcp
with IPv6 support or ucspi-ssl patched with IPv6 support which I personally
set proto to SSL or SSL6 depending on the remote ip is IPv4 or IPv6.

ex1: proto=TCP6 and name=REMOTEIP, the final fullname.s will be "TCP6REMOTEIP"
ex2: proto=SSL and name=REMOTEIP, the final fullname.s will be "SSLREMOTEIP"
ex3: proto=SSL6 and name=REMOTEIP, the final fullname.s will be "SSL6REMOTEIP"
ex4: proto=TCP6 and name=LOCALIP, the final fullname.s will be "TCP6LOCALIP"

Signed-off-by: Giam Teck Choon <giamteckchoon@gmail.com>